# idanmoosh-healthySauerKraut-1.0.0-swagger.yaml
openapi: 3.0.0
info:
  description: 'healthy sauerkraut'
  version: "1.0.0"
  title: Healthy sauerkraut
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: Idan Haim
    email: idanmoosh@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://petstore.swagger.io/v2'
...
paths:
  /auth/register:
    post:
      tags:
        - register
      summary: using user input to perform registration
      operationId: register
      responses:
        '201':
          description: Created
        '405':
          description: Invalid input
        '500':
          description: Internal Server Error
      requestBody:
        $ref: ''
  /auth/login:
    post:
      tags:
        - login
      summary: using user input to perform authentication
      operationId: loginUser
      responses:
        '200':
          description: OK
        '405':
          description: Invalid input
        '500':
          description: Internal Server Error
...